+++ src/pages/Chat/components/ConnectionStatus.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const ConnectionStatus: React.FC = () => (
+    <div className="connection-status">Disconnected</div>
+);
+
+export default ConnectionStatus;
